requirements.txt
langchain==0.0.268
streamlit==1.25.0
openai==0.27.8
google-api-python-client==2.98.0
pinecone-client==2.2.2
tiktoken==0.4.0
